+++ lib/slack/index.js
@@ -1,6 +1,6 @@
 const bodyParser = require('body-parser');
 
-const unfurl = require('./unfurl');
+const unfurl = require('../unfurls');
 const commands = require('../commands');
 const oauth = require('./oauth');
 const importer = require('./importer');
