+++ lib/unfurls/account.js
@@ -1,4 +1,4 @@
-const { Account } = require('../../messages/account');
+const { Account } = require('../messages/account');
 
 module.exports = async (params, github, unfurlType) => {
   const { owner } = params;
