+++ lib/unfurls/blob.js
@@ -1,4 +1,4 @@
-const { Blob } = require('../../messages/blob');
+const { Blob } = require('../messages/blob');
 
 module.exports = async (params, github, unfurlType) => {
   const {
