+++ lib/unfurls/comment.js
@@ -1,4 +1,4 @@
-const { Comment } = require('../../messages/comment');
+const { Comment } = require('../messages/comment');
 
 module.exports = async (params, github, unfurlType) => {
   const {
