+++ lib/unfurls/index.js
@@ -1,8 +1,8 @@
 const GitHubApi = require('probot/lib/github');
 
-const cache = require('../../cache');
+const cache = require('../cache');
 
-const githubUrl = require('../../github-url');
+const githubUrl = require('../github-url');
 
 class UnsupportedResource extends Error {
   constructor(url) {
