+++ lib/unfurls/issue.js
@@ -1,4 +1,4 @@
-const { Issue } = require('../../messages/issue');
+const { Issue } = require('../messages/issue');
 
 module.exports = async (params, github, unfurlType) => {
   const { owner, repo, number } = params;
