+++ lib/unfurls/pull.js
@@ -1,4 +1,4 @@
-const { PullRequest } = require('../../messages/pull-request');
+const { PullRequest } = require('../messages/pull-request');
 
 module.exports = async (params, github, unfurlType) => {
   const { owner, repo, number } = params;
