+++ lib/unfurls/repo.js
@@ -1,4 +1,4 @@
-const { Repository } = require('../../messages/repository');
+const { Repository } = require('../messages/repository');
 
 module.exports = async (params, github, unfurlType) => {
   const { owner, repo } = params;
